backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
scikit-learn==1.8.0
joblib==1.5.3
pydantic==2.13.3
pydantic-settings==2.14.0
python-dotenv==1.2.2
google-genai==1.73.1
httpx==0.28.1
pytest==9.0.3
